requirements.txt
colorednoise==1.1.1
matplotlib==3.1.3
rpy2==3.3.6
statsmodels==0.11.0
scipy==1.4.1
numpy==1.18.1
scikit_image==0.17.2
tqdm==4.43.0
pandas==1.0.3
PyYAML==5.4.1
scikit-image
